rake時にActiveJobの登録を行わないようにする。起動時にActiveJobを発行する都合で最初のmigrateをした時、ActiveJob記録用のテーブルが未定義で落ちるのを回避

+++ config/initializers/active_job.rb
@@ -1,8 +1,10 @@
 
 Rails.application.config.active_job.queue_adapter = :delayed_job
-Delayed::Job.destroy_all
-ArpWatchJob.perform_later
-
-at_exit do
+if "bin/rake"
   Delayed::Job.destroy_all
+  ArpWatchJob.perform_later
+
+  at_exit do
+    Delayed::Job.destroy_all
+  end
 end
